+++ module-domain-menu/src/main/java/kitchenpos/menu/MenuProductRepository.java
@@ -1,0 +1,16 @@
+package kitchenpos.menu;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
+    MenuProduct save(MenuProduct entity);
+
+    Optional<MenuProduct> findById(Long id);
+
+    List<MenuProduct> findAll();
+
+    List<MenuProduct> findAllByMenuIdIn(List<Long> menuIds);
+}
