+++ module-domain-menu/src/main/java/kitchenpos/menu/MenuRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.menu;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+    Menu save(Menu entity);
+
+    Optional<Menu> findById(Long id);
+
+    List<Menu> findAll();
+}
