+++ module-domain-menu/src/main/java/kitchenpos/menugroup/MenuGroupRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.menugroup;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
+}
