+++ module-domain-product/src/main/java/kitchenpos/product/ProductRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
