+++ settings.gradle
@@ -2,4 +2,5 @@ rootProject.name = 'kitchenpos'
 include 'module-common'
 include 'module-domain-product'
 include 'module-api'
+include 'module-domain-menu'
