--- src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,68 +1,0 @@
-package kitchenpos.menu.domain;
-
-import kitchenpos.common.domain.Price;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-@Entity
-public class Menu {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Embedded
-    private Price price;
-
-    @JoinColumn(name = "menu_group_id", foreignKey = @ForeignKey(name = "fk_menu_menu_group"), nullable = false)
-    private Long menuGroupId;
-
-    public Menu() {
-    }
-
-    public Menu(String name, Price price, Long menuGroupId) {
-        this(0L, name, price, menuGroupId);
-    }
-
-    public Menu(Long id, String name, Price price, Long menuGroupId) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.menuGroupId = menuGroupId;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Price getPrice() {
-        return price;
-    }
-
-    public Long getMenuGroupId() {
-        return menuGroupId;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Menu menu = (Menu) o;
-        return Objects.equals(id, menu.id)
-                && Objects.equals(name, menu.name)
-                && Objects.equals(price, menu.price)
-                && Objects.equals(menuGroupId, menu.menuGroupId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, price, menuGroupId);
-    }
-}
