--- src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,16 +1,0 @@
-package kitchenpos.menu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
-    MenuProduct save(MenuProduct entity);
-
-    Optional<MenuProduct> findById(Long id);
-
-    List<MenuProduct> findAll();
-
-    List<MenuProduct> findAllByMenuIdIn(List<Long> menuIds);
-}
