--- src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,14 +1,0 @@
-package kitchenpos.menu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-    Menu save(Menu entity);
-
-    Optional<Menu> findById(Long id);
-
-    List<Menu> findAll();
-}
